+++ CHANGELOG.md
@@ -6,11 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## Unreleased
 
+### Added
+
+- Added refresh of newspapers when a work is updated
+
 ## [9.2.0](https://github.com/dbmdz/cudami/releases/tag/9.2.0) – 2024-07-08
 
 ### Added
 
 - Added search slot on all pages
+
+### Fixed
+
 - Fixed delivery of error pages
 
 ## [9.1.0](https://github.com/dbmdz/cudami/releases/tag/9.1.0) – 2024-06-17